Clean up relationship tests
