--- src/steps/__tests__/__fixtures__/accountUserEntities.json
@@ -1,24 +1,0 @@
-[
-  {
-    "username": "account@gitlab.com",
-    "description": "",
-    "active": true,
-    "_key": "gitlab-account:1",
-    "_type": "gitlab_account",
-    "_class": ["User"],
-    "name": "account@gitlab.com",
-    "createdOn": 1586982455921,
-    "displayName": "account@gitlab.com"
-  },
-  {
-    "username": "user1@gitlab.com",
-    "description": "",
-    "active": true,
-    "_key": "gitlab-user:2",
-    "_type": "gitlab_user",
-    "_class": ["User"],
-    "name": "user1@gitlab.com",
-    "createdOn": 1586982455921,
-    "displayName": "user1@gitlab.com"
-  }
-]
