--- src/steps/__tests__/__fixtures__/userApprovedMergeRequestEntities.json
@@ -1,45 +1,0 @@
-{
-  "entities": [
-    {
-      "id": "gitlab-user:1",
-      "username": "user1@gitlab.com",
-      "description": "",
-      "active": true,
-      "_key": "gitlab-user:1",
-      "_type": "gitlab_user",
-      "_class": ["User"],
-      "name": "user1@gitlab.com",
-      "createdOn": 1586982455921,
-      "displayName": "user1@gitlab.com"
-    },
-    {
-      "id": "gitlab-merge-request:1",
-      "_key": "gitlab-merge-request:1",
-      "_type": "gitlab_merge_request",
-      "_class": ["PR"],
-      "name": "Merge Request 1",
-      "authorId": "gitlab-user:1",
-      "projectId": "gitlab-project:1",
-      "createdOn": 1586982455921
-    }
-  ],
-  "responses": [
-    {
-      "id": 1,
-      "iid": 1,
-      "project_id": 1,
-      "title": "title",
-      "state": "opened",
-      "description": "description",
-      "approved": true,
-      "approved_by": [
-        {
-          "user": {
-            "id": 1,
-            "name": "user1@gitlab.com"
-          }
-        }
-      ]
-    }
-  ]
-}
