--- src/steps/__tests__/__fixtures__/userProjectsEntities.json
@@ -1,22 +1,0 @@
-[
-  {
-    "id": "gitlab-user:1",
-    "username": "user1@gitlab.com",
-    "description": "",
-    "active": true,
-    "_key": "gitlab-user:1",
-    "_type": "gitlab_user",
-    "_class": ["User"],
-    "name": "user1@gitlab.com",
-    "createdOn": 1586982455921,
-    "displayName": "user1@gitlab.com"
-  },
-  {
-    "id": "gitlab-project:1",
-    "_key": "gitlab-project:1",
-    "_type": "gitlab_project",
-    "_class": ["Project"],
-    "name": "Project 1",
-    "createdOn": 1586982455921
-  }
-]
