+++ src/steps/build-group-projects-relationships/__tests__/__fixtures__/entities.json
@@ -1,0 +1,27 @@
+[
+  {
+    "id": "gitlab-group:7799803",
+    "_key": "gitlab-group:7799803",
+    "_type": "gitlab_group",
+    "_class": ["Group"],
+    "name": "CreativiceTest",
+    "createdOn": 1586982455921
+  },
+  {
+    "id": "gitlab-group:7799819",
+    "_key": "gitlab-group:7799819",
+    "_type": "gitlab_group",
+    "_class": ["Group"],
+    "name": "CreativiceSubGroup",
+    "parentGroupId": "gitlab-group:7799803",
+    "createdOn": 1586982455921
+  },
+  {
+    "id": "gitlab-project:18463260",
+    "_key": "gitlab-project:18463260",
+    "_type": "gitlab_project",
+    "_class": ["Project"],
+    "name": "CreativiceSubGroupProject",
+    "createdOn": 1586982455921
+  }
+]
