+++ src/steps/build-group-subgroup-relationships/__tests__/__fixtures__/entities.json
@@ -1,0 +1,19 @@
+[
+  {
+    "id": "gitlab-group:7799803",
+    "_key": "gitlab-group:7799803",
+    "_type": "gitlab_group",
+    "_class": ["Group"],
+    "name": "CreativiceTest",
+    "createdOn": 1586982455921
+  },
+  {
+    "id": "gitlab-group:7799819",
+    "_key": "gitlab-group:7799819",
+    "_type": "gitlab_group",
+    "_class": ["Group"],
+    "name": "CreativiceSubGroup",
+    "parentGroupId": "gitlab-group:7799803",
+    "createdOn": 1586982455921
+  }
+]
