+++ src/steps/build-group-user-relationships/__tests__/__recordings__/group_users_1316805215/recording.har
@@ -1,0 +1,243 @@
+{
+  "log": {
+    "_recordingName": "group_users",
+    "creator": {
+      "comment": "persister:JupiterOneIntegationFSPersister",
+      "name": "Polly.JS",
+      "version": "4.1.0"
+    },
+    "entries": [
+      {
+        "_id": "32098aeca30bba315903a2616a1c01b5",
+        "_order": 0,
+        "cache": {},
+        "request": {
+          "bodySize": 0,
+          "cookies": [],
+          "headers": [
+            {
+              "_fromType": "array",
+              "name": "private-token",
+              "value": "[REDACTED]"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept",
+              "value": "*/*"
+            },
+            {
+              "_fromType": "array",
+              "name": "user-agent",
+              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept-encoding",
+              "value": "gzip,deflate"
+            },
+            {
+              "_fromType": "array",
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "name": "host",
+              "value": "gitlab.com"
+            }
+          ],
+          "headersSize": 275,
+          "httpVersion": "HTTP/1.1",
+          "method": "GET",
+          "queryString": [
+            {
+              "name": "page",
+              "value": "1"
+            },
+            {
+              "name": "per_page",
+              "value": "100"
+            }
+          ],
+          "url": "https://gitlab.com/api/v4/groups/7799803/members/all?page=1&per_page=100"
+        },
+        "response": {
+          "bodySize": 549,
+          "content": {
+            "mimeType": "application/json",
+            "size": 549,
+            "text": "[\"1f8b08000000000000\",\"03ac904b6bc3301084ff4ad9b388e5876c59107aedb9f410688b594beb4420cb410fb750fadf4b1b689ad053e96d9881dd99eff10dac0125a4ec2a2918789c0914d02e059a69070c72a470e5de038398307d7aa8935d0918e08a09c39083030587948e511545249d036df6e1946ef43217275988a62b1b815c50a7498f92b76539e2c485e82b51caea366e257fca9c57add95a433e59bd7860f042e3d597bd4d0ec7afe33f2ba2d614e3e06825074a7006f47ab481e2800994cfcebdb3f3fc8657dff31f88e67c7367c3ba78f2971012d17cb0e13f10104e06bb4937533d1934bcc37a9c78df63d396ba97f43704e7829700eadf003c7f000000ffff03006418348d01020000\"]"
+          },
+          "cookies": [
+            {
+              "domain": ".gitlab.com",
+              "expires": "2020-06-11T11:04:25.000Z",
+              "httpOnly": true,
+              "name": "__cfduid",
+              "path": "/",
+              "sameSite": "Lax",
+              "secure": true,
+              "value": "[REDACTED]"
+            }
+          ],
+          "headers": [
+            {
+              "name": "date",
+              "value": "Tue, 12 May 2020 11:04:25 GMT"
+            },
+            {
+              "name": "content-type",
+              "value": "application/json"
+            },
+            {
+              "name": "transfer-encoding",
+              "value": "chunked"
+            },
+            {
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "_fromType": "array",
+              "name": "set-cookie",
+              "value": "[REDACTED]"
+            },
+            {
+              "name": "cache-control",
+              "value": "max-age=0, private, must-revalidate"
+            },
+            {
+              "name": "etag",
+              "value": "W/\"953ca7f1da525edbeec1e5c102851d70\""
+            },
+            {
+              "name": "link",
+              "value": "<https://gitlab.com/api/v4/groups/7799803/members/all?id=7799803&page=1&per_page=100>; rel=\"first\", <https://gitlab.com/api/v4/groups/7799803/members/all?id=7799803&page=1&per_page=100>; rel=\"last\""
+            },
+            {
+              "name": "vary",
+              "value": "Origin"
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-frame-options",
+              "value": "SAMEORIGIN"
+            },
+            {
+              "name": "x-next-page",
+              "value": ""
+            },
+            {
+              "name": "x-page",
+              "value": "1"
+            },
+            {
+              "name": "x-per-page",
+              "value": "100"
+            },
+            {
+              "name": "x-prev-page",
+              "value": ""
+            },
+            {
+              "name": "x-request-id",
+              "value": "gWW42Agngna"
+            },
+            {
+              "name": "x-runtime",
+              "value": "0.045306"
+            },
+            {
+              "name": "x-total",
+              "value": "2"
+            },
+            {
+              "name": "x-total-pages",
+              "value": "1"
+            },
+            {
+              "name": "strict-transport-security",
+              "value": "max-age=31536000"
+            },
+            {
+              "name": "referrer-policy",
+              "value": "strict-origin-when-cross-origin"
+            },
+            {
+              "name": "ratelimit-limit",
+              "value": "600"
+            },
+            {
+              "name": "ratelimit-observed",
+              "value": "1"
+            },
+            {
+              "name": "ratelimit-remaining",
+              "value": "599"
+            },
+            {
+              "name": "ratelimit-reset",
+              "value": "1589281525"
+            },
+            {
+              "name": "ratelimit-resettime",
+              "value": "Tue, 12 May 2020 11:05:25 GMT"
+            },
+            {
+              "name": "gitlab-lb",
+              "value": "fe-07-lb-gprd"
+            },
+            {
+              "name": "gitlab-sv",
+              "value": "localhost"
+            },
+            {
+              "name": "cf-cache-status",
+              "value": "DYNAMIC"
+            },
+            {
+              "name": "expect-ct",
+              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
+            },
+            {
+              "name": "server",
+              "value": "cloudflare"
+            },
+            {
+              "name": "cf-ray",
+              "value": "5923a8a52a83fe2c-HEL"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
+            },
+            {
+              "name": "cf-request-id",
+              "value": "02aa27bb340000fe2ca2ad7200000001"
+            }
+          ],
+          "headersSize": 1338,
+          "httpVersion": "HTTP/1.1",
+          "redirectURL": "",
+          "status": 200,
+          "statusText": "OK"
+        },
+        "startedDateTime": "2020-05-12T11:04:25.083Z",
+        "time": 270,
+        "timings": {
+          "blocked": -1,
+          "connect": -1,
+          "dns": -1,
+          "receive": 0,
+          "send": 0,
+          "ssl": -1,
+          "wait": 270
+        }
+      }
+    ],
+    "pages": [],
+    "version": "1.2"
+  }
+}
