+++ src/steps/build-project-merge-requests-relationships/__tests__/__fixtures__/entities.json
@@ -1,0 +1,19 @@
+[
+  {
+    "id": "gitlab-project:18463260",
+    "_key": "gitlab-project:18463260",
+    "_type": "gitlab_project",
+    "_class": ["Project"],
+    "name": "CreativiceSubGroupProject",
+    "createdOn": 1586982455921
+  },
+  {
+    "id": "gitlab-merge-request:18463260",
+    "_key": "gitlab-merge-request:18463260",
+    "_type": "gitlab_merge_request",
+    "_class": ["PR"],
+    "name": "Merge Request 1",
+    "projectId": "gitlab-project:18463260",
+    "createdOn": 1586982455921
+  }
+]
