+++ src/steps/build-user-approved-merge-request-relationships/__tests__/__recordings__/user_approved_merge_requests_2969860846/recording.har
@@ -1,0 +1,206 @@
+{
+  "log": {
+    "_recordingName": "user_approved_merge_requests",
+    "creator": {
+      "comment": "persister:JupiterOneIntegationFSPersister",
+      "name": "Polly.JS",
+      "version": "4.1.0"
+    },
+    "entries": [
+      {
+        "_id": "9daf3a767263b8f253d54195c48be60a",
+        "_order": 0,
+        "cache": {},
+        "request": {
+          "bodySize": 0,
+          "cookies": [],
+          "headers": [
+            {
+              "_fromType": "array",
+              "name": "private-token",
+              "value": "[REDACTED]"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept",
+              "value": "*/*"
+            },
+            {
+              "_fromType": "array",
+              "name": "user-agent",
+              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept-encoding",
+              "value": "gzip,deflate"
+            },
+            {
+              "_fromType": "array",
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "name": "host",
+              "value": "gitlab.com"
+            }
+          ],
+          "headersSize": 273,
+          "httpVersion": "HTTP/1.1",
+          "method": "GET",
+          "queryString": [],
+          "url": "https://gitlab.com/api/v4/projects/18463260/merge_requests/1/approvals"
+        },
+        "response": {
+          "bodySize": 949,
+          "content": {
+            "mimeType": "application/json",
+            "size": 949,
+            "text": "[\"1f8b08000000000000036c525d6fdb300cfc2b019f9d447622c711500cc3d6c7bd141d50ac2b045a661c6d8aed49548a21e87f1fe424cd66ec4dfcd01d79c713d80694dcacca32cfab0c6c0af30c06dfff20c37a0cab75b92a4a91015b76040abe0e0d32cd1eee3f7efe72bf3834904143c1783bb0ed3b50f0694fe6e78cf736cc4cdfbbd98e90a327c820307282e807ea287d349e90a9d1c8a0a01085988bf57c251e45a572a956db45be2abe410671e4fca74fce45f598172adf2a512c8494a9ef40be259d68620005063b5d931eb3890e87c1f7476a40b18f748dd105ede957b49ecefbdfd28e760c4a6470a9eb014378ed7da3b9d71734503b74816ee8bafe0deaf904319007753a8b5c559ba292197478480ad0137b3ad053da2d909f641ffed20a0ddb63120f8fc8e875f4\",\"0e14ec9987a096cb40267a5ab4fe5c5d98feb03c3f9772bdc9d71285a48d215357a2ccf31a7742ca6d21f3aaf810ee2af13d0a5194cd9d6da8636bfa0e3278a57ac2d25a76588fe0b711dfde5e3208b16d298cd69cb7f701d4f3cbbb18d350b7be8fc3259936d77b0c7aeacb5848e6fd5f6274da4747577b12d40de55abc629d4f22f947816f536a3ca27558a793be3446c7767034c19936befd010000ffff03002e51241336030000\"]"
+          },
+          "cookies": [
+            {
+              "domain": ".gitlab.com",
+              "expires": "2020-06-11T11:28:31.000Z",
+              "httpOnly": true,
+              "name": "__cfduid",
+              "path": "/",
+              "sameSite": "Lax",
+              "secure": true,
+              "value": "[REDACTED]"
+            }
+          ],
+          "headers": [
+            {
+              "name": "date",
+              "value": "Tue, 12 May 2020 11:28:32 GMT"
+            },
+            {
+              "name": "content-type",
+              "value": "application/json"
+            },
+            {
+              "name": "transfer-encoding",
+              "value": "chunked"
+            },
+            {
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "_fromType": "array",
+              "name": "set-cookie",
+              "value": "[REDACTED]"
+            },
+            {
+              "name": "cache-control",
+              "value": "max-age=0, private, must-revalidate"
+            },
+            {
+              "name": "etag",
+              "value": "W/\"165b377aa40c1a6c165d1d0cd61e6916\""
+            },
+            {
+              "name": "vary",
+              "value": "Origin"
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-frame-options",
+              "value": "SAMEORIGIN"
+            },
+            {
+              "name": "x-request-id",
+              "value": "SRLCOPPKJA9"
+            },
+            {
+              "name": "x-runtime",
+              "value": "0.095199"
+            },
+            {
+              "name": "strict-transport-security",
+              "value": "max-age=31536000"
+            },
+            {
+              "name": "referrer-policy",
+              "value": "strict-origin-when-cross-origin"
+            },
+            {
+              "name": "ratelimit-limit",
+              "value": "600"
+            },
+            {
+              "name": "ratelimit-observed",
+              "value": "2"
+            },
+            {
+              "name": "ratelimit-remaining",
+              "value": "598"
+            },
+            {
+              "name": "ratelimit-reset",
+              "value": "1589282972"
+            },
+            {
+              "name": "ratelimit-resettime",
+              "value": "Tue, 12 May 2020 11:29:32 GMT"
+            },
+            {
+              "name": "gitlab-lb",
+              "value": "fe-07-lb-gprd"
+            },
+            {
+              "name": "gitlab-sv",
+              "value": "localhost"
+            },
+            {
+              "name": "cf-cache-status",
+              "value": "DYNAMIC"
+            },
+            {
+              "name": "expect-ct",
+              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
+            },
+            {
+              "name": "server",
+              "value": "cloudflare"
+            },
+            {
+              "name": "cf-ray",
+              "value": "5923cbf7191dfe3c-HEL"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
+            },
+            {
+              "name": "cf-request-id",
+              "value": "02aa3dce6e0000fe3c6a2bf200000001"
+            }
+          ],
+          "headersSize": 1045,
+          "httpVersion": "HTTP/1.1",
+          "redirectURL": "",
+          "status": 200,
+          "statusText": "OK"
+        },
+        "startedDateTime": "2020-05-12T11:28:30.438Z",
+        "time": 1714,
+        "timings": {
+          "blocked": -1,
+          "connect": -1,
+          "dns": -1,
+          "receive": 0,
+          "send": 0,
+          "ssl": -1,
+          "wait": 1714
+        }
+      }
+    ],
+    "pages": [],
+    "version": "1.2"
+  }
+}
