+++ src/steps/build-user-opened-merge-request-relationships/__tests__/__fixtures__/entities.json
@@ -1,0 +1,33 @@
+[
+  {
+    "id": "gitlab-user:5887285",
+    "username": "eXtremeX",
+    "description": "",
+    "active": true,
+    "_key": "gitlab-user:5887285",
+    "_type": "gitlab_user",
+    "_class": ["User"],
+    "name": "eXtremeX",
+    "createdOn": 1586982455921,
+    "displayName": "eXtremeX"
+  },
+  {
+    "id": "gitlab-project:18463260",
+    "_key": "gitlab-project:18463260",
+    "_type": "gitlab_project",
+    "_class": ["Project"],
+    "name": "CreativiceSubGroupProject",
+    "createdOn": 1586982455921
+  },
+  {
+    "id": "gitlab-merge-request:58110137",
+    "iid": 1,
+    "authorId": "gitlab-user:5887285",
+    "_key": "gitlab-merge-request:58110137",
+    "_type": "gitlab_merge_request",
+    "_class": ["PR"],
+    "name": "Merge Request 1",
+    "projectId": "gitlab-project:18463260",
+    "createdOn": 1586982455921
+  }
+]
