+++ src/steps/build-user-projects-relationships/__tests__/__recordings__/user_projects_233881013/recording.har
@@ -1,0 +1,243 @@
+{
+  "log": {
+    "_recordingName": "user_projects",
+    "creator": {
+      "comment": "persister:JupiterOneIntegationFSPersister",
+      "name": "Polly.JS",
+      "version": "4.1.0"
+    },
+    "entries": [
+      {
+        "_id": "5e6433749e4f2fb5955c9a2438e9e433",
+        "_order": 0,
+        "cache": {},
+        "request": {
+          "bodySize": 0,
+          "cookies": [],
+          "headers": [
+            {
+              "_fromType": "array",
+              "name": "private-token",
+              "value": "[REDACTED]"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept",
+              "value": "*/*"
+            },
+            {
+              "_fromType": "array",
+              "name": "user-agent",
+              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
+            },
+            {
+              "_fromType": "array",
+              "name": "accept-encoding",
+              "value": "gzip,deflate"
+            },
+            {
+              "_fromType": "array",
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "name": "host",
+              "value": "gitlab.com"
+            }
+          ],
+          "headersSize": 278,
+          "httpVersion": "HTTP/1.1",
+          "method": "GET",
+          "queryString": [
+            {
+              "name": "page",
+              "value": "1"
+            },
+            {
+              "name": "per_page",
+              "value": "100"
+            }
+          ],
+          "url": "https://gitlab.com/api/v4/projects/18463260/members/all?page=1&per_page=100"
+        },
+        "response": {
+          "bodySize": 551,
+          "content": {
+            "mimeType": "application/json",
+            "size": 551,
+            "text": "[\"1f\",\"8b0800000000000003ac904b4b03311485ff8adc75e8641e99c9048a5bd7e2a2a052ee24376d203353f21805f1bf4bad5a2baec4dde51c38f77ce7fe059c0125a4ec1a5e3198702450704734e6ab1b179679a20918e448e1c34b44e3de05601013a6a3823ab98580012e98306c73f0a0609fd221aaa288a473a0d52e9cdc959ec7e2741684d6606775636b6bd0f00eebc1f2bec7a62d752fe93aae257fc89c57ad593b4353727a3ed679a2e1c7979d4b1e87f7f07341d49a62dc7a5ac883aa39037a3eb840718b09d494bd7f65e7012a29be06a04d0a34d2e612fd53bdfd0f78d1746523900bea34e941f2b62c07b45c88be12a5acfe06ffbde225bef80dfff10d0000ffff03007623300401020000\"]"
+          },
+          "cookies": [
+            {
+              "domain": ".gitlab.com",
+              "expires": "2020-06-11T10:36:47.000Z",
+              "httpOnly": true,
+              "name": "__cfduid",
+              "path": "/",
+              "sameSite": "Lax",
+              "secure": true,
+              "value": "[REDACTED]"
+            }
+          ],
+          "headers": [
+            {
+              "name": "date",
+              "value": "Tue, 12 May 2020 10:36:47 GMT"
+            },
+            {
+              "name": "content-type",
+              "value": "application/json"
+            },
+            {
+              "name": "transfer-encoding",
+              "value": "chunked"
+            },
+            {
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "_fromType": "array",
+              "name": "set-cookie",
+              "value": "[REDACTED]"
+            },
+            {
+              "name": "cache-control",
+              "value": "max-age=0, private, must-revalidate"
+            },
+            {
+              "name": "etag",
+              "value": "W/\"3846ac9a027429a357726a9addfdd4b2\""
+            },
+            {
+              "name": "link",
+              "value": "<https://gitlab.com/api/v4/projects/18463260/members/all?id=18463260&page=1&per_page=100>; rel=\"first\", <https://gitlab.com/api/v4/projects/18463260/members/all?id=18463260&page=1&per_page=100>; rel=\"last\""
+            },
+            {
+              "name": "vary",
+              "value": "Origin"
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-frame-options",
+              "value": "SAMEORIGIN"
+            },
+            {
+              "name": "x-next-page",
+              "value": ""
+            },
+            {
+              "name": "x-page",
+              "value": "1"
+            },
+            {
+              "name": "x-per-page",
+              "value": "100"
+            },
+            {
+              "name": "x-prev-page",
+              "value": ""
+            },
+            {
+              "name": "x-request-id",
+              "value": "ZkaDMvxSV88"
+            },
+            {
+              "name": "x-runtime",
+              "value": "0.090352"
+            },
+            {
+              "name": "x-total",
+              "value": "2"
+            },
+            {
+              "name": "x-total-pages",
+              "value": "1"
+            },
+            {
+              "name": "strict-transport-security",
+              "value": "max-age=31536000"
+            },
+            {
+              "name": "referrer-policy",
+              "value": "strict-origin-when-cross-origin"
+            },
+            {
+              "name": "ratelimit-limit",
+              "value": "600"
+            },
+            {
+              "name": "ratelimit-observed",
+              "value": "1"
+            },
+            {
+              "name": "ratelimit-remaining",
+              "value": "599"
+            },
+            {
+              "name": "ratelimit-reset",
+              "value": "1589279867"
+            },
+            {
+              "name": "ratelimit-resettime",
+              "value": "Tue, 12 May 2020 10:37:47 GMT"
+            },
+            {
+              "name": "gitlab-lb",
+              "value": "fe-06-lb-gprd"
+            },
+            {
+              "name": "gitlab-sv",
+              "value": "localhost"
+            },
+            {
+              "name": "cf-cache-status",
+              "value": "DYNAMIC"
+            },
+            {
+              "name": "expect-ct",
+              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
+            },
+            {
+              "name": "server",
+              "value": "cloudflare"
+            },
+            {
+              "name": "cf-ray",
+              "value": "5923802bdb89fe24-HEL"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
+            },
+            {
+              "name": "cf-request-id",
+              "value": "02aa0e6f620000fe2420995200000001"
+            }
+          ],
+          "headersSize": 1346,
+          "httpVersion": "HTTP/1.1",
+          "redirectURL": "",
+          "status": 200,
+          "statusText": "OK"
+        },
+        "startedDateTime": "2020-05-12T10:36:47.306Z",
+        "time": 281,
+        "timings": {
+          "blocked": -1,
+          "connect": -1,
+          "dns": -1,
+          "receive": 0,
+          "send": 0,
+          "ssl": -1,
+          "wait": 281
+        }
+      }
+    ],
+    "pages": [],
+    "version": "1.2"
+  }
+}
